agents/programming/win10-11/requirements.txt
langchain
python-dotenv
requests
openai
